refactor(frontend): data stratification
Organise data stratification into a `src/features/data/stratification` folder.
